Add files via upload
This version of the bot has been reworked to use locally hosted MySQL database. In addition the bot utilize COGS, which is a way to separate the commands by groups into separate files. Those files will be loaded into the "Main" file when launching the bot.

PS: Any reference made to any IDs or config. file are from a file that was not provided due to security purposes.